Waveshare2p8 (#86)
* Waveshare esp32-s3 2.8

https://www.waveshare.com/esp32-s3-touch-lcd-2.8.htm
requires my drivers for the cst328

* Update esp32-s3touchlcd2p8.json

Update for the latest version of smartdisplay with new values.
There is some very small artifacting related to the DMA but it is hardly noticable. I cannot figure out the perfect values for DMA so someone else may want to look into this in the future. The performance benefit of having dma vs disabling it is significant though, so i will take the minimal artifacting for now over laggy responsiveness when not using dma.